fix a small windows compilation warning
git-svn-id: http://svn.cern.ch/guest/lhcb/Rec/trunk@21919 4525493e-7705-40b1-a816-d608a930855b

Former-commit-id: d414adc46a7dbced7ed308db331505c24c0f6405

+++ Rec/Muon/MuonID/src/MuonIDAlg.cpp
@@ -1500,7 +1500,7 @@ double MuonIDAlg::calc_ProbNonMu(const double& dist0, const double *parNonMu){
 // date:    10/5/09
 //=====================================================================
 
-  float x = m_x;     // x-width for the integral
+  float x = (float)(m_x);     // x-width for the integral
   int nMax = m_nMax;   // number of steps  
   double landau = 0;
   double Prob=0;
